requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
appier==1.21.7
async-timeout==3.0.1
asyncpraw==7.2.0
asyncprawcore==2.0.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
discord.py==1.7.2
emoji==1.2.0
greenlet==1.1.0
idna==2.10
joblib==1.0.1
lxml==4.6.3
multidict==5.1.0
networkx==2.5.1
patsy==0.5.1
praw==7.2.0
prawcore==2.0.0
pydot==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
update-checker==0.18.0
urllib3==1.26.4
websocket-client==1.0.0
youtube-dlc==2020.11.11.post3

